move ResultV1 to test file
