+++ src/globus_compute_common/messagepack/protocol.py
@@ -7,6 +7,7 @@ A MessagePackProtocol only needs to define two actions:
 
 And the two must be inverses on any valid inputs.
 """
+
 import abc
 
 from .message_types import Message
